adds typings to remaining files
