--- src/utils/sleep.js
@@ -1,10 +1,0 @@
-module.exports = {
-  /** @type {(ms: number) => Promise<void>} */
-  sleep(ms) {
-    return new Promise((resolve) => {
-      setTimeout(() => {
-        resolve();
-      }, ms);
-    });
-  },
-};
