+++ tsconfig.json
@@ -1,13 +1,14 @@
 {
   "compilerOptions": {
     "strict": true,
-    "allowJs": true,
+    "allowJs": false,
     "noEmit": true,
     "allowSyntheticDefaultImports": true,
     "lib": ["ES2020"],
     "module": "ES2020",
     "moduleResolution": "node",
-    "target": "ES2020"
+    "target": "ES2020",
+    "typeRoots": ["**/*.d.ts"]
   },
   "include": ["src", "typings.d.ts"]
 }
